use system contracts
- replace "localnet" with "emulator" in network flag help
- fix filename const in filenames.go
