+++ model/bootstrap/filenames.go
@@ -20,7 +20,7 @@ var (
 	PathNodeInfosPub          = filepath.Join(DirnamePublicBootstrap, "node-infos.pub.json")
 	PathPartnerNodeInfoPrefix = filepath.Join(DirnamePublicBootstrap, "node-info.pub.")
 	PathNodeInfoPub           = filepath.Join(DirnamePublicBootstrap, "node-info.pub.%v.json") // %v will be replaced by NodeID
-	PartnerStakesFileName     = "partner-stakes.json"
+	FileNamePartnerStakes     = "partner-stakes.json"
 	DirnameRootBlockVotes     = filepath.Join(DirnamePublicBootstrap, "root-block-votes")
 
 	PathRootBlockData             = filepath.Join(DirnamePublicBootstrap, "root-block.json")
